修复：通知页 SMTP 授权码提交格式
