--- internal/webui/static/assets/Chat-8cCRP4iO.js
@@ -1,1 +1,0 @@
-import{d as E,j as e,a}from"./react-vendor-Bf9ELcll.js";import{L as pe}from"./vendor-BE-_xUHH.js";import{B as me}from"./audio-codec-mjVr_wwJ.js";import{L as G,an as de,ao as ge,u as xe,ap as ue,aq as he,M as Q0,V as _e,R as be,ag as D0,m as fe,ar as je,x as J0,as as we,at as Ne,au as ye,av as Ce,S as W0,X as Oe}from"./icons-vendor-BSp7PTZg.js";import{a as U0,i as R,d as i0,e as V0,f as z0,m as Z0,h as ve,j as ke,k as q0,l as Se,s as Re,n as Ee,o as Ie,p as Me,g as Le,q as G0,u as $0,t as P0,C as De}from"./chat-metadata-74T0w35U.js";import{p as Ue,s as qe}from"./index-dO7JyNQY.js";const H0="ydisks:chat-audio-play",Pe=[6,10,15,8,17,12,7,16,11,6,14,9];function Te(u){const _=u.trim();if(typeof location>"u"||location.protocol!=="https:")return _;try{const n=new URL(_);return n.protocol==="http:"&&n.hostname.endsWith(".aliyuncs.com")&&(n.protocol="https:"),n.toString()}catch{return _}}function T0(u){const _=Number.isFinite(u)?Math.max(0,Math.floor(u)):0,n=Math.floor(_/60),l=_%60;return`${n}:${l.toString().padStart(2,"0")}`}function A0(u){window.dispatchEvent(new CustomEvent(H0,{detail:u}))}const Ae=({messageKey:u,src:_,outgoing:n,initialDuration:l=0})=>{const x=E.useRef(null),C=E.useRef(0),[o,I]=E.useState("idle"),[M,L]=E.useState(l),[V,T]=E.useState(0);E.useEffect(()=>(()=>{var d;C.current+=1,(d=x.current)==null||d.destroy(),x.current=null}),[_]),E.useEffect(()=>{L(Math.max(0,l))},[l]),E.useEffect(()=>{const d=f=>{const w=f.detail,N=x.current;w!==u&&(N!=null&&N.isPlaying())&&N.pause()};return window.addEventListener(H0,d),(()=>window.removeEventListener(H0,d))},[u]),E.useEffect(()=>{if(o!=="playing")return;const d=window.setInterval(()=>{const f=x.current;f&&T(f.getCurrentPosition())},100);return(()=>window.clearInterval(d))},[o]);const A=(d,f)=>{const w=()=>C.current===f&&x.current===d;d.onPlay(()=>{w()&&I("playing")}),d.onResume(()=>{w()&&I("playing")}),d.onPause(()=>{w()&&I("paused")}),d.onStop(()=>{w()&&(I("idle"),T(0))}),d.onEnded(()=>{w()&&(I("idle"),T(0))})},r0=async()=>{var w;const d=x.current;if(d!=null&&d.isPlaying()){d.pause();return}if(d!=null&&d.isPaused()){A0(u),d.resume();return}if(d!=null&&d.isInit()){A0(u),d.play();return}if(o==="loading")return;const f=C.current+1;C.current=f,I("loading"),T(0);try{const N=new me;if(x.current=N,A(N,f),await N.initWithUrl(Te(_)),C.current!==f||x.current!==N){N.destroy();return}L(N.getDuration()||Math.max(0,l)),A0(u),N.play()}catch{if(C.current!==f)return;(w=x.current)==null||w.destroy(),x.current=null,I("error"),T(0)}},H=o==="playing"||o==="paused"?V:M,$=M>0?o==="playing"||o==="paused"?`${T0(H)} / ${T0(M)}`:T0(M):"--:--",P=o==="loading"?"正在解码":o==="playing"?"正在播放":o==="paused"?"已暂停":o==="error"?"加载失败，点击重试":"语音消息",O=o==="playing"?"暂停语音消息":o==="paused"?"继续播放语音消息":o==="error"?"重新加载语音消息":o==="loading"?"正在加载语音消息":"播放语音消息";return e.jsxs("div",{className:`flex h-11 min-w-56 items-center gap-2 rounded-2xl border px-2 shadow-sm ${n?"rounded-br-md border-sky-400 bg-sky-500 text-white":"rounded-bl-md border-slate-200 bg-white text-slate-800"}`,children:[e.jsx("button",{type:"button","aria-label":O,title:O,disabled:o==="loading",onClick:(()=>void r0()),className:`flex h-7 w-7 shrink-0 items-center justify-center rounded-full transition focus:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 disabled:cursor-wait ${n?"bg-white text-sky-600 hover:bg-sky-50 focus-visible:ring-white":"bg-sky-50 text-sky-600 hover:bg-sky-100 focus-visible:ring-sky-400"}`,children:o==="loading"?e.jsx(G,{className:"h-4 w-4 animate-spin"}):o==="playing"?e.jsx(de,{className:"h-4 w-4 fill-current"}):o==="error"?e.jsx(ge,{className:"h-4 w-4"}):e.jsx(xe,{className:"ml-px h-4 w-4 fill-current"})}),e.jsxs("div",{className:"min-w-0 flex flex-1 items-center gap-2",children:[e.jsx("div",{className:"flex h-5 flex-1 items-center gap-0.5","aria-hidden":"true",children:Pe.map((d,f)=>e.jsx("span",{className:`w-0.5 rounded-full ${o==="playing"?"animate-pulse":""} ${n?"bg-sky-100":o==="error"?"bg-red-300":"bg-sky-400"}`,style:{height:`${d}px`,animationDelay:`${f*45}ms`}},`${u}-${f}`))}),e.jsx("span",{className:`shrink-0 font-mono text-[11px] font-semibold tabular-nums ${n?"text-sky-100":o==="error"?"text-red-500":"text-slate-500"}`,children:$}),e.jsx("span",{className:"sr-only",children:P})]})]})},X0=[["尊嘟假嘟","https://img.alicdn.com/imgextra/i2/O1CN01hWzOgD1KcmFEIaGwE_!!6000000001185-2-tps-64-64.png_110x10000.jpg_.webp"],["鸡哔你","https://img.alicdn.com/imgextra/i1/O1CN01P9TVJl1JCqCZcQIdW_!!6000000000993-2-tps-64-64.png_110x10000.jpg_.webp"],["谁懂啊","https://img.alicdn.com/imgextra/i1/O1CN01r9qUnx1g3tcrtSmcs_!!6000000004087-2-tps-64-64.png_110x10000.jpg_.webp"],["吗喽","https://img.alicdn.com/imgextra/i4/O1CN01dTynZh1LnbIYTHI1D_!!6000000001344-2-tps-64-64.png_110x10000.jpg_.webp"],["功德+1","https://img.alicdn.com/imgextra/i1/O1CN01l1G5pS1IVn37pfQJi_!!6000000000899-2-tps-64-64.png_110x10000.jpg_.webp"],["city不city啊","https://img.alicdn.com/imgextra/i2/O1CN01zRjPTR29lWHqabmDA_!!6000000008108-2-tps-64-64.png_110x10000.jpg_.webp"],["激推","https://img.alicdn.com/imgextra/i3/O1CN01ADTJ6u1khypfxbfnQ_!!6000000004716-2-tps-64-64.png_110x10000.jpg_.webp"],["微笑","https://img.alicdn.com/imgextra/i1/O1CN017b61R51wRk2lVV7vx_!!6000000006305-2-tps-980-980.png_110x10000.jpg_.webp"],["龇牙","https://img.alicdn.com/imgextra/i2/O1CN01nmOh4S22Yxz8IdEVf_!!6000000007133-2-tps-980-980.png_110x10000.jpg_.webp"],["流泪","https://img.alicdn.com/imgextra/i2/O1CN01Ec1EaL1zqQETcM2dX_!!6000000006765-2-tps-980-980.png_110x10000.jpg_.webp"],["尴尬","https://img.alicdn.com/imgextra/i2/O1CN01UrWH7a1ZVBrgRYXlc_!!6000000003199-2-tps-980-980.png_110x10000.jpg_.webp"],["撇嘴","https://img.alicdn.com/imgextra/i2/O1CN01Sr7o2b1lUWg5KfOD8_!!6000000004822-2-tps-980-980.png_110x10000.jpg_.webp"],["破涕为笑","https://img.alicdn.com/imgextra/i4/O1CN01pPetah1gkwndfn7Fs_!!6000000004181-2-tps-980-980.png_110x10000.jpg_.webp"],["捂脸哭","https://img.alicdn.com/imgextra/i2/O1CN0170KgpE1dE1TuQgd66_!!6000000003703-2-tps-980-980.png_110x10000.jpg_.webp"],["害羞","https://img.alicdn.com/imgextra/i3/O1CN01ZoB0Lb1ps7XpqP8PQ_!!6000000005415-2-tps-980-980.png_110x10000.jpg_.webp"],["感动","https://img.alicdn.com/imgextra/i2/O1CN01wqD4ly29yndR3Qp3A_!!6000000008137-2-tps-980-980.png_110x10000.jpg_.webp"],["滴泪","https://img.alicdn.com/imgextra/i1/O1CN01MpkJ3L1oUTTqfryOn_!!6000000005228-2-tps-980-980.png_110x10000.jpg_.webp"],["暗中观察","https://img.alicdn.com/imgextra/i1/O1CN010LO9yj272W07GlLls_!!6000000007739-2-tps-980-980.png_110x10000.jpg_.webp"],["偷笑","https://img.alicdn.com/imgextra/i4/O1CN01vJtO031g818IuYO7x_!!6000000004096-2-tps-980-980.png_110x10000.jpg_.webp"],["ok","https://img.alicdn.com/imgextra/i3/O1CN01QRkqDF25tz0WcANje_!!6000000007585-2-tps-980-980.png_110x10000.jpg_.webp"],["送花","https://img.alicdn.com/imgextra/i2/O1CN01PpzBNX1JNpeJBWnvr_!!6000000001017-2-tps-980-980.png_110x10000.jpg_.webp"],["大哭","https://img.alicdn.com/imgextra/i4/O1CN014TiaNb1ViEgNvRrTB_!!6000000002686-2-tps-980-980.png_110x10000.jpg_.webp"],["色","https://img.alicdn.com/imgextra/i2/O1CN01840t1B1JaCLGvHnBK_!!6000000001044-2-tps-980-980.png_110x10000.jpg_.webp"],["惊慌","https://img.alicdn.com/imgextra/i1/O1CN01mnqO7G1LAfUnfAN2U_!!6000000001259-2-tps-980-980.png_110x10000.jpg_.webp"],["惊讶","https://img.alicdn.com/imgextra/i3/O1CN0127IeyS1Kw0p0P4jzI_!!6000000001227-2-tps-980-980.png_110x10000.jpg_.webp"],["得意","https://img.alicdn.com/imgextra/i4/O1CN018QUnFv1WpOUES9POZ_!!6000000002837-2-tps-980-980.png_110x10000.jpg_.webp"],["睡","https://img.alicdn.com/imgextra/i2/O1CN01RprGZl1lqVgnEJLCx_!!6000000004870-2-tps-980-980.png_110x10000.jpg_.webp"],["不好意思","https://img.alicdn.com/imgextra/i3/O1CN01rMJJgn1SUXzuVpUhO_!!6000000002250-2-tps-980-980.png_110x10000.jpg_.webp"],["托腮","https://img.alicdn.com/imgextra/i1/O1CN01w3reBU1d3wlgOBtyZ_!!6000000003681-2-tps-980-980.png_110x10000.jpg_.webp"],["发怒","https://img.alicdn.com/imgextra/i2/O1CN013Wzy8d2A8PxGJZPpT_!!6000000008158-2-tps-980-980.png_110x10000.jpg_.webp"],["难过","https://img.alicdn.com/imgextra/i2/O1CN01zcwaSh1G9gZCFuFal_!!6000000000580-2-tps-980-980.png_110x10000.jpg_.webp"],["调皮","https://img.alicdn.com/imgextra/i3/O1CN01Zd1YHi1TI0lwpBe8c_!!6000000002358-2-tps-980-980.png_110x10000.jpg_.webp"],["愉快","https://img.alicdn.com/imgextra/i2/O1CN0145wDk21SWpyjTvsvA_!!6000000002255-2-tps-980-980.png_110x10000.jpg_.webp"],["白眼","https://img.alicdn.com/imgextra/i4/O1CN01nexGth1hwEAT1SHO2_!!6000000004341-2-tps-980-980.png_110x10000.jpg_.webp"],["拒绝","https://img.alicdn.com/imgextra/i2/O1CN01nqusqS1uHXzoUcpXr_!!6000000006012-2-tps-980-980.png_110x10000.jpg_.webp"],["期待","https://img.alicdn.com/imgextra/i3/O1CN01M44QY11Oh8Q5PdMIx_!!6000000001736-2-tps-980-980.png_110x10000.jpg_.webp"],["困","https://img.alicdn.com/imgextra/i4/O1CN01J4hFFz1Y9pmNdBq76_!!6000000003017-2-tps-980-980.png_110x10000.jpg_.webp"],["流汗","https://img.alicdn.com/imgextra/i4/O1CN01BU3NZ21gdcSjkqnST_!!6000000004165-2-tps-980-980.png_110x10000.jpg_.webp"],["憨笑","https://img.alicdn.com/imgextra/i2/O1CN01Pee9Ql1SJ661eH9By_!!6000000002225-2-tps-980-980.png_110x10000.jpg_.webp"],["耶","https://img.alicdn.com/imgextra/i2/O1CN01pFicsL1ipBhIqzcgs_!!6000000004461-2-tps-980-980.png_110x10000.jpg_.webp"],["奋斗","https://img.alicdn.com/imgextra/i3/O1CN01ZhQ4AV25xBmTK3CnF_!!6000000007592-2-tps-980-980.png_110x10000.jpg_.webp"],["咒骂","https://img.alicdn.com/imgextra/i3/O1CN01r1Zh9927QKb4pnf7k_!!6000000007791-2-tps-980-980.png_110x10000.jpg_.webp"],["疑问","https://img.alicdn.com/imgextra/i4/O1CN01UZqbzt29MmlNctCQS_!!6000000008054-2-tps-980-980.png_110x10000.jpg_.webp"],["嘘","https://img.alicdn.com/imgextra/i1/O1CN01qBlYkM1ojaYF75On5_!!6000000005261-2-tps-980-980.png_110x10000.jpg_.webp"],["晕","https://img.alicdn.com/imgextra/i2/O1CN014zhM2h1wPuSUFoMwV_!!6000000006301-2-tps-980-980.png_110x10000.jpg_.webp"],["开心到模糊","https://img.alicdn.com/imgextra/i1/O1CN01T1Dy271j6akZWo7Vr_!!6000000004499-2-tps-980-980.png_110x10000.jpg_.webp"],["衰","https://img.alicdn.com/imgextra/i1/O1CN01aOXsNr1LmgMvtyzOO_!!6000000001342-2-tps-980-980.png_110x10000.jpg_.webp"],["委屈","https://img.alicdn.com/imgextra/i2/O1CN013J0bMk1QNYih3UFTk_!!6000000001964-2-tps-980-980.png_110x10000.jpg_.webp"],["再见","https://img.alicdn.com/imgextra/i2/O1CN01IHKsQx1UdpFgeTHcA_!!6000000002541-2-tps-980-980.png_110x10000.jpg_.webp"],["鼓掌","https://img.alicdn.com/imgextra/i4/O1CN01DHxB5s1G1RRBxekIu_!!6000000000562-2-tps-980-980.png_110x10000.jpg_.webp"],["吐","https://img.alicdn.com/imgextra/i1/O1CN01xyVxH21LNUaMgW4hh_!!6000000001287-2-tps-980-980.png_110x10000.jpg_.webp"],["坏笑","https://img.alicdn.com/imgextra/i1/O1CN01zV7opv1bIU6iI8OhM_!!6000000003442-2-tps-980-980.png_110x10000.jpg_.webp"],["抠鼻","https://img.alicdn.com/imgextra/i1/O1CN01QOanju1sNqCrtYstu_!!6000000005755-2-tps-980-980.png_110x10000.jpg_.webp"],["叹气","https://img.alicdn.com/imgextra/i3/O1CN01RElx7R1QD1c8ZN07f_!!6000000001941-2-tps-980-980.png_110x10000.jpg_.webp"],["无聊","https://img.alicdn.com/imgextra/i2/O1CN01d91LnZ1UJDQjKCouC_!!6000000002496-2-tps-980-980.png_110x10000.jpg_.webp"],["思考","https://img.alicdn.com/imgextra/i2/O1CN01wQjg2f1tO84ZWitn9_!!6000000005891-2-tps-980-980.png_110x10000.jpg_.webp"],["呆","https://img.alicdn.com/imgextra/i1/O1CN01GLqBWC1DGbhqmpD2L_!!6000000000189-2-tps-980-980.png_110x10000.jpg_.webp"],["哇哦","https://img.alicdn.com/imgextra/i2/O1CN01MtLpS21qRO2bnAA5k_!!6000000005492-2-tps-980-980.png_110x10000.jpg_.webp"],["飞吻","https://img.alicdn.com/imgextra/i4/O1CN01q1WpFE25ljsNKjvHX_!!6000000007567-2-tps-980-980.png_110x10000.jpg_.webp"],["好喜欢","https://img.alicdn.com/imgextra/i2/O1CN01mD4iQM1O57XyQ5SHD_!!6000000001653-2-tps-980-980.png_110x10000.jpg_.webp"],["拜托","https://img.alicdn.com/imgextra/i4/O1CN01SQgrTH1dzCH3pIyRN_!!6000000003806-2-tps-980-980.png_110x10000.jpg_.webp"],["真棒","https://img.alicdn.com/imgextra/i2/O1CN017fcAmd1SCgYCj9Upk_!!6000000002211-2-tps-980-980.png_110x10000.jpg_.webp"],["流鼻血","https://img.alicdn.com/imgextra/i3/O1CN0139JscY1rC6RS9GR4j_!!6000000005594-2-tps-980-980.png_110x10000.jpg_.webp"],["费解","https://img.alicdn.com/imgextra/i1/O1CN01FqAVvR1tyJMPj13ie_!!6000000005970-2-tps-980-980.png_110x10000.jpg_.webp"],["秃头","https://img.alicdn.com/imgextra/i4/O1CN01ZF4bV11thMhqm2Z1h_!!6000000005933-2-tps-980-980.png_110x10000.jpg_.webp"],["嘿哈","https://img.alicdn.com/imgextra/i4/O1CN01ObcOw61FOVlvQg7Mw_!!6000000000477-2-tps-980-980.png_110x10000.jpg_.webp"],["吃土","https://img.alicdn.com/imgextra/i3/O1CN01Z6nJ1b1ZYr22YseHN_!!6000000003207-2-tps-980-980.png_110x10000.jpg_.webp"],["机智","https://img.alicdn.com/imgextra/i1/O1CN01rfKY351zDUZJlBiir_!!6000000006680-2-tps-980-980.png_110x10000.jpg_.webp"],["举杯","https://img.alicdn.com/imgextra/i4/O1CN01R2tn4F1mn8NuEIE8J_!!6000000004998-2-tps-980-980.png_110x10000.jpg_.webp"],["握手","https://img.alicdn.com/imgextra/i4/O1CN01NDcA251JDkvuFhSDb_!!6000000000995-2-tps-980-980.png_110x10000.jpg_.webp"],["挥手","https://img.alicdn.com/imgextra/i4/O1CN01iVN0ht1TmEuSwauGt_!!6000000002424-2-tps-980-980.png_110x10000.jpg_.webp"],["yeah","https://img.alicdn.com/imgextra/i3/O1CN01opt7hh27Wk8omjR4t_!!6000000007805-2-tps-980-980.png_110x10000.jpg_.webp"],["666","https://img.alicdn.com/imgextra/i3/O1CN01YRWDPF1OjQOpXeuTC_!!6000000001741-2-tps-980-980.png_110x10000.jpg_.webp"],["棒","https://img.alicdn.com/imgextra/i3/O1CN01agprvs1HU7rPRxOnB_!!6000000000760-2-tps-980-980.png_110x10000.jpg_.webp"],["爱你","https://img.alicdn.com/imgextra/i1/O1CN01SzxgcH1GSSo6Lwmwk_!!6000000000621-2-tps-980-980.png_110x10000.jpg_.webp"],["比心","https://img.alicdn.com/imgextra/i1/O1CN01YHZ39t1DeQIeb0TWQ_!!6000000000241-2-tps-980-980.png_110x10000.jpg_.webp"],["感谢","https://img.alicdn.com/imgextra/i3/O1CN01ywacF81baLYNeyK1n_!!6000000003481-2-tps-980-980.png_110x10000.jpg_.webp"],["拍手","https://img.alicdn.com/imgextra/i4/O1CN01wbj4wv1K7ArD9AU44_!!6000000001116-2-tps-980-980.png_110x10000.jpg_.webp"],["拉钩","https://img.alicdn.com/imgextra/i1/O1CN01dePkKH1HOd6ivAcu8_!!6000000000748-2-tps-980-980.png_110x10000.jpg_.webp"],["没钱","https://img.alicdn.com/imgextra/i3/O1CN01qhxGjp1zD2AfqmWx5_!!6000000006679-2-tps-980-980.png_110x10000.jpg_.webp"],["小刀","https://img.alicdn.com/imgextra/i3/O1CN01tFVx5Y1T2thPjhR6b_!!6000000002325-2-tps-980-980.png_110x10000.jpg_.webp"],["不刀","https://img.alicdn.com/imgextra/i4/O1CN01zTxff61clcvOcTS3n_!!6000000003641-2-tps-980-980.png_110x10000.jpg_.webp"],["同出","https://img.alicdn.com/imgextra/i3/O1CN01URKas6249R8ozS98O_!!6000000007348-2-tps-980-980.png_110x10000.jpg_.webp"],["递钱","https://img.alicdn.com/imgextra/i4/O1CN01PzhgV51FLlOaZHZ6n_!!6000000000471-2-tps-980-980.png_110x10000.jpg_.webp"],["数钱","https://img.alicdn.com/imgextra/i3/O1CN01CzXXvp23khkljToQH_!!6000000007294-2-tps-980-980.png_110x10000.jpg_.webp"],["打call","https://img.alicdn.com/imgextra/i1/O1CN01X7VUVm1il48IYv8Qw_!!6000000004452-2-tps-980-980.png_110x10000.jpg_.webp"],["号外","https://img.alicdn.com/imgextra/i2/O1CN01vg6xSM1Vxo9EKyfjR_!!6000000002720-2-tps-980-980.png_110x10000.jpg_.webp"],["包邮","https://img.alicdn.com/imgextra/i3/O1CN01yVS7ys24jA22oVZ2T_!!6000000007426-2-tps-980-980.png_110x10000.jpg_.webp"],["求赞","https://img.alicdn.com/imgextra/i1/O1CN01Xwqi5A1giCQMRcX6N_!!6000000004175-2-tps-980-980.png_110x10000.jpg_.webp"],["非主流","https://img.alicdn.com/imgextra/i1/O1CN013wyI1j1hlEfDUoATI_!!6000000004317-2-tps-980-980.png_110x10000.jpg_.webp"],["吃瓜","https://img.alicdn.com/imgextra/i1/O1CN01LLOFsh1UUCw0deZba_!!6000000002520-2-tps-980-980.png_110x10000.jpg_.webp"],["种草","https://img.alicdn.com/imgextra/i2/O1CN01ooeRC81hA8a4x9S3q_!!6000000004236-2-tps-980-980.png_110x10000.jpg_.webp"],["佛系","https://img.alicdn.com/imgextra/i4/O1CN01BYXEe321GMHMZ9w66_!!6000000006957-2-tps-980-980.png_110x10000.jpg_.webp"],["超便宜","https://img.alicdn.com/imgextra/i4/O1CN01mqgSZi1Er4rOIDiP7_!!6000000000404-2-tps-980-980.png_110x10000.jpg_.webp"],["拥抱","https://img.alicdn.com/imgextra/i1/O1CN01yuGD9c1hJkttfKHAm_!!6000000004257-2-tps-980-980.png_110x10000.jpg_.webp"],["摊手","https://img.alicdn.com/imgextra/i2/O1CN014AiIqE1o9repE6STR_!!6000000005183-2-tps-980-980.png_110x10000.jpg_.webp"],["举手","https://img.alicdn.com/imgextra/i1/O1CN01xizuHZ1WPkJCogbuj_!!6000000002781-2-tps-980-980.png_110x10000.jpg_.webp"],["蹲","https://img.alicdn.com/imgextra/i3/O1CN01Midjjn1FrMj2CWer6_!!6000000000540-2-tps-980-980.png_110x10000.jpg_.webp"],["跪了","https://img.alicdn.com/imgextra/i3/O1CN01F2y52j24E16L6RIjF_!!6000000007358-2-tps-980-980.png_110x10000.jpg_.webp"],["拜","https://img.alicdn.com/imgextra/i4/O1CN01Jldd1G1oXDrEHH2qr_!!6000000005234-2-tps-980-980.png_110x10000.jpg_.webp"],["排","https://img.alicdn.com/imgextra/i3/O1CN01uSUmYD1jD0IJf0UVU_!!6000000004513-2-tps-980-980.png_110x10000.jpg_.webp"],["旺财","https://img.alicdn.com/imgextra/i3/O1CN01A25Mv01DKjGseAmuP_!!6000000000198-2-tps-980-980.png_110x10000.jpg_.webp"],["招财猫","https://img.alicdn.com/imgextra/i3/O1CN01WmJ7Sr20EhDyp86Bd_!!6000000006818-2-tps-980-980.png_110x10000.jpg_.webp"],["酸了","https://img.alicdn.com/imgextra/i3/O1CN01UFgjn61JJi5Iqxdk7_!!6000000001008-2-tps-980-980.png_110x10000.jpg_.webp"],["鸭鸭","https://img.alicdn.com/imgextra/i3/O1CN01pHSCSQ1deaS8WDiqG_!!6000000003761-2-tps-980-980.png_110x10000.jpg_.webp"],["方了","https://img.alicdn.com/imgextra/i3/O1CN01br7g711whlu8xOPUt_!!6000000006340-2-tps-980-980.png_110x10000.jpg_.webp"],["白菜","https://img.alicdn.com/imgextra/i1/O1CN01KABFLt24WKwXjVp08_!!6000000007398-2-tps-980-980.png_110x10000.jpg_.webp"],["桃子","https://img.alicdn.com/imgextra/i1/O1CN01CQhhPH1vJfRjDGBv5_!!6000000006152-2-tps-980-980.png_110x10000.jpg_.webp"],["打包","https://img.alicdn.com/imgextra/i3/O1CN01oYOkUI1VT7c0ivZHV_!!6000000002653-2-tps-980-980.png_110x10000.jpg_.webp"],["沙发","https://img.alicdn.com/imgextra/i4/O1CN018qUFaE1DGbhqpMIit_!!6000000000189-2-tps-980-980.png_110x10000.jpg_.webp"],["南","https://img.alicdn.com/imgextra/i1/O1CN01ZSA6wg1PEZKf4h99R_!!6000000001809-2-tps-980-980.png_110x10000.jpg_.webp"],["盯","https://img.alicdn.com/imgextra/i4/O1CN01OWB1Gx1s01bzBhHN6_!!6000000005703-2-tps-980-980.png_110x10000.jpg_.webp"],["hot","https://img.alicdn.com/imgextra/i4/O1CN01DDD9uA22cAl7VDH7d_!!6000000007140-2-tps-980-980.png_110x10000.jpg_.webp"],["new","https://img.alicdn.com/imgextra/i3/O1CN01rRJXJo1qwzN4YnOHX_!!6000000005561-2-tps-980-980.png_110x10000.jpg_.webp"],["cool","https://img.alicdn.com/imgextra/i4/O1CN013bUubF2855qb2Qt5l_!!6000000007880-2-tps-980-980.png_110x10000.jpg_.webp"],["1","https://img.alicdn.com/imgextra/i1/O1CN01U8XG5y1Ie22frQHCQ_!!6000000000917-2-tps-980-980.png_110x10000.jpg_.webp"],["可","https://img.alicdn.com/imgextra/i2/O1CN01SMjtPG1iJaN17p94N_!!6000000004392-2-tps-980-980.png_110x10000.jpg_.webp"],["搬砖","https://img.alicdn.com/imgextra/i1/O1CN01GbqW6G1LOrm0Gllcb_!!6000000001290-2-tps-980-980.png_110x10000.jpg_.webp"],["菜狗","https://img.alicdn.com/imgextra/i2/O1CN01VtN0t91USNLmvfmol_!!6000000002516-2-tps-980-980.png_110x10000.jpg_.webp"],["开心面具","https://img.alicdn.com/imgextra/i3/O1CN017elkpk22EoYlBPKd3_!!6000000007089-2-tps-980-980.png_110x10000.jpg_.webp"],["快递","https://img.alicdn.com/imgextra/i2/O1CN01AmUZXk1d3wle5O8IJ_!!6000000003681-2-tps-980-980.png_110x10000.jpg_.webp"],["裂开","https://img.alicdn.com/imgextra/i1/O1CN01BihwvO1a5N9bFt0LU_!!6000000003278-2-tps-980-980.png_110x10000.jpg_.webp"],["啤酒","https://img.alicdn.com/imgextra/i4/O1CN01jDkgEj1WvLdXUVBDx_!!6000000002850-2-tps-980-980.png_110x10000.jpg_.webp"],["瑞思拜","https://img.alicdn.com/imgextra/i4/O1CN011gR0Um1K3VgpYjPOi_!!6000000001108-2-tps-980-980.png_110x10000.jpg_.webp"],["小丑","https://img.alicdn.com/imgextra/i3/O1CN01jMikhO1NJUM15XKTf_!!6000000001549-2-tps-980-980.png_110x10000.jpg_.webp"],["爱心","https://img.alicdn.com/imgextra/i1/O1CN015RVSWf23s25eETJqO_!!6000000007310-2-tps-980-980.png_110x10000.jpg_.webp"]],He=u=>u,Fe=new Map(X0.map(([u,_])=>[u,_])),Be=u=>u.split(/(\[[^\]]+\])/g).map((_,n)=>{const l=_.startsWith("[")?Fe.get(_.slice(1,-1)):void 0;return l?e.jsx("img",{src:l,alt:_,title:_,className:"mx-0.5 inline-block h-7 w-7 align-middle"},n):e.jsx(E.Fragment,{children:_},n)}),Qe=()=>{const[u,_]=a.useState([]),[n,l]=a.useState(""),[x,C]=a.useState({}),[o,I]=a.useState(""),[M,L]=a.useState([]),[V,T]=a.useState(""),[A,r0]=a.useState(!1),[H,$]=a.useState(""),[P,O]=a.useState(null),[d,f]=a.useState(!0),[w,N]=a.useState(!1),[X,j0]=a.useState(!1),[l0,w0]=a.useState(!1),[N0,c0]=a.useState(),[o0,p0]=a.useState({}),[m0,d0]=a.useState({}),[g0,x0]=a.useState(!1),[S0,R0]=a.useState(!1),[v,Y]=a.useState(!1),[u0,k]=a.useState(""),[E0,I0]=a.useState("connecting"),[K,y0]=a.useState(null),[z,C0]=a.useState(null),Z=a.useRef(""),F=a.useRef(""),e0=a.useRef(null),t0=a.useRef({accountID:"",chatID:""}),s0=a.useRef(!0),J=a.useRef(!1),U=a.useRef(null),h0=a.useRef(new Set),s=a.useRef(0),h=a.useRef(null),S=a.useRef(0),B=a.useRef(null),q=a.useRef(0),_0=a.useRef(null),b0=a.useRef(0),O0=a.useRef(null),Q=a.useRef(0),a0=a.useRef(null);a.useEffect(()=>{Z.current=n},[n]),a.useEffect(()=>{F.current=o},[o]);const f0=a.useCallback(async t=>{var p;const i=++s.current;(p=h.current)==null||p.abort();const c=new AbortController;h.current=c;try{const r=await U0(t,void 0,{signal:c.signal},!0);if(!R(s.current,i,c.signal))return[];const g=r.sessions.map(m=>t===Z.current&&m.chat_id===F.current?{...m,unread_count:0}:m);return C(m=>({...m,[t]:g})),p0(m=>({...m,[t]:r.next_cursor})),d0(m=>({...m,[t]:r.has_more})),g}catch(r){return R(s.current,i,c.signal)&&!i0(r)&&k(r instanceof Error?r.message:"同步会话失败"),[]}},[]);a.useEffect(()=>{const t=new AbortController;return(async()=>{var c;f(!0);try{const[p,r]=await Promise.all([Le({signal:t.signal}),G0({signal:t.signal})]),m=p.map(y=>{var W,B0;return{...y,runtime_state:((W=r[y.id])==null?void 0:W.state)||(y.enabled?"connecting":"disabled"),runtime_connected:((B0=r[y.id])==null?void 0:B0.connected)===!0}}).filter(y=>y.enabled),j=await Promise.all(m.map(async y=>[y.id,await U0(y.id,void 0,{signal:t.signal})]));if(t.signal.aborted)return;_(m),C(Object.fromEntries(j.map(([y,W])=>[y,W.sessions]))),p0(Object.fromEntries(j.map(([y,W])=>[y,W.next_cursor]))),d0(Object.fromEntries(j.map(([y,W])=>[y,W.has_more])));const D=window.localStorage.getItem("ydisks.chat.account.v1")||"",L0=m.some(y=>y.id===D)?D:((c=m[0])==null?void 0:c.id)||"";l(L0)}catch(p){t.signal.aborted||k(p instanceof Error?p.message:"加载聊天数据失败")}finally{t.signal.aborted||f(!1)}})(),(()=>t.abort())},[]),a.useEffect(()=>{let t=!1,i=0,c=null;const p=async()=>{c=new AbortController;try{const r=await G0({signal:c.signal,timeoutMs:1e4});t||_(g=>g.map(m=>{var j,D;return{...m,runtime_state:((j=r[m.id])==null?void 0:j.state)||m.runtime_state,runtime_connected:((D=r[m.id])==null?void 0:D.connected)??m.runtime_connected}}))}catch{}finally{t||(i=window.setTimeout(p,3e3))}};return i=window.setTimeout(p,3e3),(()=>{t=!0,window.clearTimeout(i),c==null||c.abort()})},[]),a.useEffect(()=>{if(!n)return;window.localStorage.setItem("ydisks.chat.account.v1",n);const t=x[n]||[];I(i=>{var c;return t.some(p=>p.chat_id===i)?i:((c=t[0])==null?void 0:c.chat_id)||""})},[n,x]),a.useEffect(()=>{const t=Object.keys(x).length;if(d||u.length>0&&t===0)return;const i=Object.values(x).some(c=>c.some(p=>p.unread_count>0));Ue(i)},[u.length,d,x]),a.useEffect(()=>{!n||h0.current.has(n)||(h0.current.add(n),f0(n))},[n,f0]),a.useEffect(()=>{var c,p;if(!n||!o){S.current+=1,q.current+=1,(c=_0.current)==null||c.abort(),J.current=!1,L([]);return}const t=++S.current;(p=B.current)==null||p.abort();const i=new AbortController;return B.current=i,N(!0),V0(n,o,void 0,void 0,{signal:i.signal}).then(r=>{if(!R(S.current,t,i.signal))return;const g=z0(r.messages,o);Z0(n,o,g,{signal:i.signal}),L(r.messages),w0(r.has_more),c0(r.next_cursor),r.session&&C(m=>({...m,[n]:(m[n]||[]).map(j=>{var D;return j.chat_id===((D=r.session)==null?void 0:D.chat_id)?r.session:j})})),C(m=>({...m,[n]:(m[n]||[]).map(j=>j.chat_id===o?{...j,unread_count:0}:j)}))}).catch(r=>{R(S.current,t,i.signal)&&!i0(r)&&k(r instanceof Error?r.message:"加载消息失败")}).finally(()=>{R(S.current,t,i.signal)&&N(!1)}),(()=>i.abort())},[n,o]),a.useEffect(()=>{var t;(t=O0.current)==null||t.abort(),b0.current+=1},[n]),a.useEffect(()=>{var t;(t=a0.current)==null||t.abort(),Q.current+=1},[n,o]),a.useEffect(()=>{O(null)},[n,o]),a.useEffect(()=>{var t;q.current+=1,(t=_0.current)==null||t.abort(),J.current=!1},[n,o]),a.useEffect(()=>(()=>{var t,i,c,p,r;(t=h.current)==null||t.abort(),(i=B.current)==null||i.abort(),q.current+=1,(c=_0.current)==null||c.abort(),(p=O0.current)==null||p.abort(),(r=a0.current)==null||r.abort()}),[]),a.useEffect(()=>{const t=P;return(()=>{t&&URL.revokeObjectURL(t.url)})},[P]);const Y0=a.useCallback(async()=>{var g,m;if(!n||!o||X||!l0)return;const t=e0.current,i=(t==null?void 0:t.scrollHeight)||0,c=S.current,p=++q.current;(g=_0.current)==null||g.abort();const r=new AbortController;_0.current=r,J.current=!0,j0(!0),k("");try{const j=(m=M[0])==null?void 0:m.id,D=await V0(n,o,N0,j,{signal:r.signal});if(!R(S.current,c,r.signal)||p!==q.current)return;L(L0=>ve(L0,D.messages)),w0(D.has_more),c0(D.next_cursor),requestAnimationFrame(()=>{p===q.current&&!r.signal.aborted&&t&&(t.scrollTop+=t.scrollHeight-i)})}catch(j){J.current=!1,i0(j)||k(j instanceof Error?j.message:"加载历史消息失败")}finally{p===q.current&&j0(!1)}},[n,o,l0,N0,M,X]),F0=a.useCallback(t=>{const i=t.account_id;if(C(c=>{const p=c[i]||[];return p.some(g=>g.chat_id===t.chat_id)?{...c,[i]:p.map(g=>g.chat_id===t.chat_id?{...g,last_message:t.content,last_message_at:t.sent_at,unread_count:t.direction==="incoming"&&t.message_type!=="system"&&(Z.current!==i||F.current!==t.chat_id)?g.unread_count+1:g.unread_count}:g).sort((g,m)=>m.last_message_at-g.last_message_at)}:(f0(i),c)}),Z.current===i&&F.current===t.chat_id&&(L(c=>ke(q0(c,t),t)),t.direction==="incoming"&&t.message_type!=="system")){const c=z0([t],t.chat_id);Z0(i,t.chat_id,c)}},[f0]);a.useEffect(()=>qe(i=>{if(i.type==="connection"){I0(i.state);return}F0(i.message)}),[F0]);const K0=a.useCallback(()=>{const t=e0.current;if(!t)return;const i=t.scrollHeight-t.scrollTop-t.clientHeight;s0.current=i<=48},[]);a.useLayoutEffect(()=>{const t=t0.current.accountID!==n||t0.current.chatID!==o;t0.current={accountID:n,chatID:o},t&&(s0.current=!0);const i=e0.current;if(i){if(J.current){J.current=!1;return}(w||s0.current)&&(i.scrollTop=i.scrollHeight)}},[n,o,M,w]);const ee=u.find(t=>t.id===n),v0=x[n]||[],b=v0.find(t=>t.chat_id===o),te=a.useMemo(()=>Se(v0,V,A),[v0,V,A]),se=a.useCallback(t=>(x[t]||[]).reduce((i,c)=>i+c.unread_count,0),[x]),ae=a.useCallback(async()=>{var p;if(!n||g0||!m0[n])return;const t=++b0.current;(p=O0.current)==null||p.abort();const i=new AbortController;O0.current=i;const c=n;x0(!0),k("");try{const r=await U0(c,o0[c],{signal:i.signal},!0);if(!R(b0.current,t,i.signal))return;C(g=>({...g,[c]:r.sessions})),p0(g=>({...g,[c]:r.next_cursor})),d0(g=>({...g,[c]:r.has_more}))}catch(r){R(b0.current,t,i.signal)&&!i0(r)&&k(r instanceof Error?r.message:"加载历史联系人失败")}finally{R(b0.current,t,i.signal)&&x0(!1)}},[n,o0,g0,m0]),n0=a.useCallback(async(t,i,c)=>{var g;if(!b||!n||v)return;const p=++Q.current;(g=a0.current)==null||g.abort();const r=new AbortController;a0.current=r,Y(!0),k("");try{const m=await Re({account_id:n,chat_id:b.chat_id,buyer_id:b.buyer_id,buyer_name:b.buyer_name,item_id:b.item_id,item_title:b.item_title,text:t},{signal:r.signal});if(!R(Q.current,p,r.signal))return;c&&$(""),y0(null),L(j=>q0(j,m.message))}catch(m){R(Q.current,p,r.signal)&&(i&&y0(t),i0(m)||k(m instanceof Error?m.message:"消息发送失败"))}finally{R(Q.current,p,r.signal)&&Y(!1)}},[n,b,v]),k0=a.useCallback(async(t,i)=>{var r;if(!b||!n||v)return;const c=++Q.current;(r=a0.current)==null||r.abort();const p=new AbortController;a0.current=p,Y(!0),k("");try{const g=await Ee({account_id:n,chat_id:b.chat_id,buyer_id:b.buyer_id,buyer_name:b.buyer_name,buyer_avatar_url:b.buyer_avatar_url,item_id:b.item_id,item_title:b.item_title,image:t},{signal:p.signal});if(!R(Q.current,c,p.signal))return;C0(null),L(m=>q0(m,g.message))}catch(g){R(Q.current,c,p.signal)&&(i&&C0(t),i0(g)||k(g instanceof Error?g.message:"图片发送失败"))}finally{R(Q.current,c,p.signal)&&(Y(!1),U.current&&(U.current.value=""))}},[n,b,v]),ne=a.useCallback(async()=>{const t=H.trim();!t||!b||!n||v||await n0(t,!0,!0)},[n,H,b,n0,v]),ie=a.useCallback(async t=>{const i=t.trim();!i||!b||!n||v||await n0(i,!0,!1)},[n,b,n0,v]),M0=a.useCallback(async t=>{if(!(!t||!b||!n||v)){if(!t.type.startsWith("image/")){k("仅支持粘贴/发送图片文件");return}k(""),O({file:t,url:URL.createObjectURL(t)})}},[n,b,v]),re=a.useCallback(async t=>{const i=t.find(c=>c.type.startsWith("image/"));i&&await M0(i)},[M0]),le=a.useCallback(()=>{O(null),U.current&&(U.current.value="")},[]),ce=a.useCallback(async()=>{if(!P||!b||!n||v)return;const t=P.file;O(null),await k0(t,!0)},[n,P,b,k0,v]),oe=a.useCallback(async()=>{K?await n0(K,!1,!0):z&&await k0(z,!1)},[z,K,k0,n0]);return{accounts:u,activeAccountID:n,activeSessions:v0,selectedSession:b,activeAccount:ee,messages:M,search:V,unreadOnly:A,draft:H,loading:d,messagesLoading:w,olderLoading:X,hasOlder:l0,contactsLoading:g0,hasMoreContacts:m0[n]===!0,emojiOpen:S0,sending:v,error:u0,liveState:E0,pendingImage:P,activeChatID:o,filteredSessions:te,scrollRef:e0,imageInputRef:U,setActiveAccountID:l,setActiveChatID:I,setSearch:T,setUnreadOnly:r0,setDraft:$,setEmojiOpen:R0,reloadSessions:f0,loadMoreContacts:ae,loadOlderMessages:Y0,handleMessageScroll:K0,handleSend:ne,handleQuickReply:ie,handleImage:M0,handlePastedImages:re,confirmSendImage:ce,closeImagePreview:le,retrySend:oe,retryAvailable:!!(K||z),unreadForAccount:se,emojiURL:He,xianyuEmojis:X0,renderXianyuText:Be,formatClock:Me,messageTime:Ie}},$e=()=>{const{accounts:u,activeAccountID:_,activeChatID:n,activeAccount:l,selectedSession:x,filteredSessions:C,messages:o,search:I,unreadOnly:M,draft:L,loading:V,messagesLoading:T,olderLoading:A,hasOlder:r0,contactsLoading:H,hasMoreContacts:$,emojiOpen:P,sending:O,error:d,liveState:f,pendingImage:w,scrollRef:N,imageInputRef:X,setActiveAccountID:j0,setActiveChatID:l0,setSearch:w0,setUnreadOnly:N0,setDraft:c0,setEmojiOpen:o0,reloadSessions:p0,loadMoreContacts:m0,loadOlderMessages:d0,handleMessageScroll:g0,handleSend:x0,handleQuickReply:S0,handleImage:R0,handlePastedImages:v,confirmSendImage:Y,closeImagePreview:u0,retrySend:k,retryAvailable:E0,unreadForAccount:I0,emojiURL:K,xianyuEmojis:y0,renderXianyuText:z,formatClock:C0,messageTime:Z}=Qe(),F=E.useMemo(()=>o.filter(s=>s.message_type==="image"),[o]),e0=E.useMemo(()=>F.map(s=>({src:s.content,alt:"聊天图片"})),[F]),[t0,s0]=E.useState(-1),J=E.useCallback(s=>{s0(F.findIndex(h=>h.message_key===s))},[F]),[U,h0]=E.useState(!1);return V?e.jsx("div",{className:"flex h-[calc(100vh-4rem)] items-center justify-center",children:e.jsx(G,{className:"h-8 w-8 animate-spin text-sky-500"})}):e.jsxs("section",{className:"flex h-[calc(100vh-4rem)] min-h-[560px] flex-col overflow-hidden rounded-2xl border border-slate-200 bg-white shadow-chat",children:[e.jsxs("header",{className:"border-b border-slate-200 bg-slate-50/70 px-5 pt-4",children:[e.jsxs("div",{className:"mb-3 flex items-center justify-between gap-4",children:[e.jsxs("div",{children:[e.jsx("h2",{className:"text-xl font-black tracking-tight text-slate-950",children:"在线聊天"}),e.jsx("p",{className:"mt-0.5 text-xs font-medium text-slate-500",children:"复用账号实时连接，消息按账号完全隔离"})]}),e.jsxs("div",{className:`flex items-center gap-2 rounded-full px-3 py-1.5 text-xs font-bold ${f==="online"?"bg-emerald-50 text-emerald-700":f==="connecting"?"bg-amber-50 text-amber-700":"bg-red-50 text-red-700"}`,children:[f==="online"?e.jsx(ue,{className:"h-3.5 w-3.5"}):e.jsx(he,{className:"h-3.5 w-3.5"}),f==="online"?"实时同步中":f==="connecting"?"正在连接":"连接已断开"]})]}),e.jsx("div",{className:"flex gap-1 overflow-x-auto pb-0",role:"tablist","aria-label":"聊天账号",children:u.map(s=>{const h=s.id===_,S=I0(s.id),B=P0(S),q=s.runtime_state==="online";return e.jsxs("button",{type:"button",role:"tab","aria-selected":h,onClick:(()=>j0(s.id)),className:`relative flex h-11 shrink-0 items-center gap-2 border-b-2 px-3 text-sm font-extrabold transition-colors ${h?"border-sky-500 text-sky-700":"border-transparent text-slate-500 hover:text-slate-900"}`,children:[e.jsx("span",{className:`h-2 w-2 rounded-full ${q?"bg-emerald-500":"bg-slate-300"}`}),e.jsx("span",{className:"max-w-36 truncate",children:s.nickname||s.remark||s.id}),S>0&&e.jsx("span",{"aria-label":`未读消息 ${B} 条`,className:$0(S),children:B})]},s.id)})})]}),u.length===0?e.jsxs("div",{className:"flex flex-1 flex-col items-center justify-center text-center",children:[e.jsx(Q0,{className:"h-12 w-12 text-slate-300"}),e.jsx("h3",{className:"mt-4 font-black text-slate-800",children:"暂无启用账号"}),e.jsx("p",{className:"mt-1 text-sm text-slate-500",children:"先在账号管理中启用账号，聊天会话会自动出现。"})]}):e.jsxs("div",{className:"grid min-h-0 flex-1 overflow-hidden grid-cols-[320px_minmax(0,1fr)]",children:[e.jsxs("aside",{className:"flex min-h-0 flex-col border-r border-slate-200 bg-slate-50/40",children:[e.jsxs("div",{className:"space-y-3 border-b border-slate-200 p-3",children:[e.jsxs("div",{className:"relative",children:[e.jsx(_e,{className:"absolute left-3 top-1/2 h-4 w-4 -translate-y-1/2 text-slate-400"}),e.jsx("input",{value:I,onChange:(s=>w0(s.target.value)),placeholder:"搜索用户、商品或消息",className:"h-10 w-full rounded-xl border border-slate-200 bg-white pl-9 pr-3 text-sm outline-none transition focus:border-sky-400 focus:ring-2 focus:ring-sky-100"})]}),e.jsxs("div",{className:"flex items-center justify-between",children:[e.jsx("button",{type:"button",onClick:(()=>N0(s=>!s)),className:`rounded-lg px-2.5 py-1.5 text-xs font-bold ${M?"bg-sky-100 text-sky-700":"text-slate-500 hover:bg-slate-100"}`,children:M?"只看未读":"全部会话"}),e.jsx("button",{type:"button",title:"刷新会话",onClick:(()=>void p0(_)),className:"rounded-lg p-2 text-slate-400 hover:bg-slate-100 hover:text-slate-700",children:e.jsx(be,{className:"h-4 w-4"})})]})]}),e.jsxs("div",{className:"min-h-0 flex-1 overflow-y-auto",children:[C.map(s=>e.jsxs("button",{type:"button",onClick:(()=>l0(s.chat_id)),className:`flex w-full gap-3 border-b border-slate-100 p-3.5 text-left transition-colors ${s.chat_id===n?"bg-white shadow-chat-active":"hover:bg-white/80"}`,children:[e.jsx("div",{className:"flex h-10 w-10 shrink-0 items-center justify-center overflow-hidden rounded-full bg-slate-200 text-slate-500",children:s.buyer_avatar_url?e.jsx("img",{src:s.buyer_avatar_url,alt:"",className:"h-full w-full object-cover"}):e.jsx(D0,{className:"h-5 w-5"})}),e.jsxs("div",{className:"min-w-0 flex-1",children:[e.jsx("div",{className:"flex items-center gap-2",children:e.jsx("span",{className:"truncate text-sm font-extrabold text-slate-900",children:s.buyer_name||`用户 ${s.buyer_id}`})}),e.jsxs("div",{className:"mt-1 flex items-center gap-2",children:[e.jsx("span",{className:"truncate text-xs text-slate-500",children:s.last_message||"暂无消息"}),s.unread_count>0&&e.jsx("span",{"aria-label":`未读消息 ${P0(s.unread_count)} 条`,className:`ml-auto ${$0(s.unread_count)}`,children:P0(s.unread_count)})]}),s.item_title&&e.jsxs("div",{className:"mt-1.5 truncate text-[10px] font-medium text-sky-700",children:["商品 · ",s.item_title]})]}),e.jsxs("div",{className:"flex shrink-0 flex-col items-end gap-1.5",children:[e.jsx("span",{className:"text-[10px] font-medium text-slate-400",children:C0(s.last_message_at)}),s.item_image_url&&e.jsx("img",{src:s.item_image_url,alt:"",className:"h-9 w-11 rounded-[4px] border border-slate-200 object-cover"})]})]},s.chat_id)),C.length===0&&e.jsx("div",{className:"px-6 py-16 text-center text-sm text-slate-400",children:"当前账号暂无匹配会话"}),$&&!I&&!M&&e.jsx("div",{className:"flex justify-center p-4",children:e.jsxs("button",{type:"button",onClick:(()=>void m0()),disabled:H,className:"flex items-center gap-2 rounded-full border border-slate-200 bg-white px-4 py-2 text-xs font-bold text-slate-500 shadow-sm hover:border-sky-200 hover:text-sky-600 disabled:opacity-50",children:[H&&e.jsx(G,{className:"h-3.5 w-3.5 animate-spin"}),H?"正在加载":"加载更多历史联系人"]})})]})]}),e.jsxs("main",{className:"relative flex min-h-0 min-w-0 flex-col overflow-hidden bg-surface-subtle",children:[x?e.jsxs(e.Fragment,{children:[e.jsxs("div",{className:"flex h-20 shrink-0 items-start border-b border-slate-200 bg-white px-5 pt-4",children:[e.jsxs("div",{className:"min-w-0",children:[e.jsx("div",{className:"truncate text-sm font-black text-slate-950",children:x.buyer_name||x.buyer_id}),e.jsxs("div",{className:"mt-0.5 flex flex-col text-xs text-slate-500",children:[e.jsx("span",{children:"用户 ID："}),e.jsx("span",{className:"truncate",children:x.buyer_id})]})]}),e.jsx("span",{className:`ml-auto rounded-full px-2.5 py-1 text-[10px] font-bold ${(l==null?void 0:l.runtime_state)==="online"?"bg-emerald-50 text-emerald-700":"bg-slate-100 text-slate-500"}`,children:(l==null?void 0:l.runtime_state)==="online"?"账号在线":"账号离线"})]}),e.jsx("div",{ref:N,onScroll:g0,className:"min-h-0 flex-1 space-y-5 overflow-y-auto px-6 py-5",children:T?e.jsx("div",{className:"flex justify-center py-12",children:e.jsx(G,{className:"h-6 w-6 animate-spin text-sky-500"})}):e.jsxs(e.Fragment,{children:[r0&&e.jsx("div",{className:"flex justify-center pb-1",children:e.jsxs("button",{type:"button",onClick:(()=>void d0()),disabled:A,className:"flex items-center gap-2 rounded-full border border-slate-200 bg-white px-4 py-1.5 text-xs font-bold text-slate-500 shadow-sm transition hover:border-sky-200 hover:text-sky-600 disabled:opacity-50",children:[A&&e.jsx(G,{className:"h-3.5 w-3.5 animate-spin"}),A?"正在加载":"加载更早消息"]})}),o.map(s=>{const h=s.direction==="outgoing";return s.message_type==="system"?e.jsx("div",{className:"flex justify-center py-1",children:e.jsxs("div",{className:"max-w-[82%] rounded-xl border border-slate-200 bg-slate-100 px-4 py-2 text-center text-xs leading-5 text-slate-500",children:[z(s.content),e.jsx("div",{className:"mt-1 text-[10px] text-slate-400",children:Z(s.sent_at)})]})},s.message_key):e.jsxs("div",{className:`flex items-end gap-2.5 ${h?"justify-end":"justify-start"}`,children:[!h&&e.jsx("div",{className:"h-9 w-9 shrink-0 overflow-hidden rounded-full bg-slate-200 ring-2 ring-white",children:x.buyer_avatar_url?e.jsx("img",{src:x.buyer_avatar_url,alt:x.buyer_name||"用户",className:"h-full w-full object-cover"}):e.jsx(D0,{className:"m-2 h-5 w-5 text-slate-500"})}),e.jsxs("div",{className:`max-w-[72%] ${h?"items-end":"items-start"} flex flex-col`,children:[e.jsx("div",{className:"mb-1 px-1 text-[10px] font-semibold text-slate-400",children:h?(l==null?void 0:l.nickname)||(l==null?void 0:l.remark)||"我":x.buyer_name||s.sender_name||x.buyer_id}),s.message_type==="image"?e.jsx("button",{type:"button",title:"点击预览大图",onClick:J.bind(null,s.message_key),className:`block cursor-zoom-in overflow-hidden rounded-2xl border bg-white p-1 text-left shadow-sm ${h?"rounded-br-md border-sky-200":"rounded-bl-md border-slate-200"}`,children:e.jsx("img",{src:s.content,alt:"聊天图片",className:"max-h-80 max-w-full rounded-xl object-contain"})}):s.message_type==="video"?e.jsx("video",{src:s.content,controls:!0,preload:"metadata",className:"max-h-80 max-w-full rounded-2xl bg-black"}):s.message_type==="audio"?e.jsx(Ae,{messageKey:s.message_key,src:s.content,outgoing:h,initialDuration:s.media_duration}):e.jsx("div",{className:`rounded-2xl px-4 py-2.5 text-sm leading-6 shadow-sm ${h?"rounded-br-md bg-sky-500 text-white":"rounded-bl-md border border-slate-200 bg-white text-slate-800"}`,children:z(s.content)}),e.jsxs("div",{className:"mt-1 flex items-center gap-1 text-[10px] text-slate-400",children:[Z(s.sent_at),h&&(s.status==="failed"?e.jsx(fe,{className:"h-3 w-3 text-red-500","aria-label":"发送失败"}):s.read_status===2?e.jsx(je,{className:"h-3 w-3 text-sky-500","aria-label":"对方已读"}):s.status==="sent"?e.jsx(J0,{className:"h-3 w-3 text-sky-500","aria-label":"已发送未读"}):e.jsx(J0,{className:"h-3 w-3","aria-label":"发送中"}))]})]}),h&&e.jsx("div",{className:"h-9 w-9 shrink-0 overflow-hidden rounded-full bg-sky-100 ring-2 ring-white",children:l!=null&&l.avatar_url?e.jsx("img",{src:l.avatar_url,alt:"我",className:"h-full w-full object-cover"}):e.jsx(D0,{className:"m-2 h-5 w-5 text-sky-600"})})]},s.message_key)})]})}),d&&e.jsxs("div",{className:"flex items-center justify-between gap-3 border-t border-red-100 bg-red-50 px-5 py-2 text-xs font-medium text-red-700",children:[e.jsx("span",{children:d}),E0&&e.jsx("button",{type:"button",className:"font-bold underline",onClick:(()=>void k()),children:"重试发送"})]}),e.jsxs("div",{className:"relative z-10 shrink-0 border-t border-slate-200 bg-white p-4 shadow-chat-input",children:[e.jsxs("div",{className:"mb-2 flex items-center gap-1",children:[e.jsxs("div",{className:"relative",children:[e.jsx("button",{type:"button",onClick:(()=>o0(s=>!s)),disabled:O||(l==null?void 0:l.runtime_state)!=="online",className:"rounded-lg p-2 text-slate-500 hover:bg-sky-50 hover:text-sky-600 disabled:opacity-40",title:"闲鱼表情",children:e.jsx(we,{className:"h-5 w-5"})}),P&&e.jsxs("div",{className:"absolute bottom-11 left-0 z-30 w-[360px] rounded-2xl border border-slate-200 bg-white p-3 shadow-2xl",children:[e.jsx("div",{className:"mb-2 text-xs font-bold text-slate-500",children:"全部表情"}),e.jsx("div",{className:"grid max-h-72 grid-cols-8 gap-1 overflow-y-auto",children:y0.map(([s,h])=>e.jsx("button",{type:"button",title:`[${s}]`,onClick:(()=>{c0(S=>S+`[${s}]`),o0(!1)}),className:"flex h-10 w-10 items-center justify-center rounded-lg hover:bg-slate-100",children:e.jsx("img",{src:K(h),alt:`[${s}]`,className:"h-8 w-8 object-contain"})},s))})]})]}),e.jsx("input",{ref:X,type:"file",accept:"image/*",className:"hidden",onChange:(s=>{var h;return void R0((h=s.target.files)==null?void 0:h[0])})}),e.jsx("button",{type:"button",onClick:(()=>{var s;return(s=X.current)==null?void 0:s.click()}),disabled:O||(l==null?void 0:l.runtime_state)!=="online",className:"rounded-lg p-2 text-slate-500 transition hover:bg-sky-50 hover:text-sky-600 disabled:opacity-40",title:"发送图片（最大 10MB）",children:e.jsx(Ne,{className:"h-5 w-5"})}),e.jsx("button",{type:"button",onClick:(()=>h0(!U)),className:"ml-auto rounded-lg p-2 text-slate-500 transition hover:bg-sky-50 hover:text-sky-600",title:U?"收起快捷回复":"展开快捷回复","aria-label":U?"收起快捷回复":"展开快捷回复",children:U?e.jsx(ye,{className:"h-5 w-5"}):e.jsx(Ce,{className:"h-5 w-5"})})]}),e.jsxs("div",{className:"flex items-end gap-3 rounded-2xl border border-slate-200 bg-slate-50 p-2 transition focus-within:border-sky-400 focus-within:ring-2 focus-within:ring-sky-100",children:[e.jsx("textarea",{value:L,onChange:(s=>c0(s.target.value)),rows:2,maxLength:2e3,onKeyDown:(s=>{s.key==="Enter"&&!s.shiftKey&&(s.preventDefault(),x0())}),onPaste:(s=>{var B;const h=Array.from(((B=s.clipboardData)==null?void 0:B.files)||[]);h.find(q=>q.type.startsWith("image/"))&&(s.preventDefault(),v(h))}),disabled:(l==null?void 0:l.runtime_state)!=="online",placeholder:(l==null?void 0:l.runtime_state)==="online"?"输入消息，Enter 发送，Shift + Enter 换行，Ctrl + V 粘贴图片":"账号离线，暂时无法发送",className:"max-h-32 min-h-12 flex-1 resize-none bg-transparent px-2 py-2 text-sm leading-6 outline-none disabled:cursor-not-allowed"}),e.jsx("button",{type:"button",onClick:(()=>void x0()),disabled:!L.trim()||O||(l==null?void 0:l.runtime_state)!=="online",className:"flex h-10 w-10 shrink-0 items-center justify-center rounded-xl bg-sky-500 text-white shadow-md shadow-sky-100 transition hover:bg-sky-600 disabled:cursor-not-allowed disabled:bg-slate-300 disabled:shadow-none","aria-label":"发送消息",children:O?e.jsx(G,{className:"h-4 w-4 animate-spin"}):e.jsx(W0,{className:"h-4 w-4"})})]})]})]}):e.jsxs("div",{className:"flex flex-1 flex-col items-center justify-center text-center",children:[e.jsx(Q0,{className:"h-12 w-12 text-slate-300"}),e.jsx("h3",{className:"mt-4 font-black text-slate-700",children:"选择一个用户开始聊天"}),e.jsx("p",{className:"mt-1 text-sm text-slate-400",children:"该账号的新消息会实时出现在左侧列表。"})]}),e.jsx(De,{activeAccountID:_,selectedSession:x,quickReplyPanelOpen:U,closeQuickReplyPanel:(()=>h0(!1)),sendQuickReply:S0,sending:O,accountOnline:(l==null?void 0:l.runtime_state)==="online"})]})]}),w&&e.jsx("div",{className:"fixed inset-0 z-50 flex items-center justify-center bg-slate-950/60 p-4",onClick:u0,children:e.jsxs("div",{className:"flex max-h-full w-full max-w-2xl flex-col overflow-hidden rounded-2xl bg-white shadow-2xl",onClick:(s=>s.stopPropagation()),children:[e.jsxs("div",{className:"flex shrink-0 items-center justify-between border-b border-slate-200 px-5 py-3",children:[e.jsx("div",{className:"text-sm font-black text-slate-900",children:"发送图片预览"}),e.jsx("button",{type:"button",title:"取消",onClick:u0,className:"rounded-lg p-1.5 text-slate-400 transition hover:bg-slate-100 hover:text-slate-700",children:e.jsx(Oe,{className:"h-5 w-5"})})]}),e.jsx("div",{className:"flex min-h-0 flex-1 items-center justify-center overflow-hidden bg-slate-100 p-4",children:e.jsx("img",{src:w.url,alt:"待发送图片预览",className:"max-h-[55vh] max-w-full rounded-xl object-contain"})}),e.jsxs("div",{className:"flex shrink-0 items-center justify-between gap-3 border-t border-slate-200 px-5 py-3",children:[e.jsxs("div",{className:"min-w-0 truncate text-xs text-slate-500",children:[w.file.name||"粘贴的图片"," · ",(w.file.size/1024).toFixed(0)," KB"]}),e.jsxs("div",{className:"flex shrink-0 items-center gap-2",children:[e.jsx("button",{type:"button",onClick:u0,className:"rounded-xl px-4 py-2 text-sm font-bold text-slate-600 transition hover:bg-slate-100",children:"取消"}),e.jsxs("button",{type:"button",onClick:(()=>void Y()),disabled:O||(l==null?void 0:l.runtime_state)!=="online",className:"flex items-center gap-2 rounded-xl bg-sky-500 px-4 py-2 text-sm font-bold text-white shadow-md shadow-sky-100 transition hover:bg-sky-600 disabled:cursor-not-allowed disabled:opacity-50",children:[O?e.jsx(G,{className:"h-4 w-4 animate-spin"}):e.jsx(W0,{className:"h-4 w-4"}),"发送"]})]})]})]})}),e.jsx(pe,{open:t0>=0,index:Math.max(t0,0),close:(()=>s0(-1)),slides:e0})]})};export{$e as default};
