+++ internal/webui/static/assets/index-Q3nlx6gR.js
@@ -1,0 +1,2 @@
+const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/Dashboard-DBUkwLYQ.js","assets/react-vendor-Bf9ELcll.js","assets/dateTime-DDi7rqpW.js","assets/icons-vendor-BSp7PTZg.js","assets/charts-vendor-DBu_rVrn.js","assets/vendor-BE-_xUHH.js","assets/vendor-DUblxfbA.css","assets/chat-metadata-74T0w35U.js","assets/contract-COFcZjyb.js","assets/AccountList-D1PDvKjr.js","assets/OrderList-CKY0um79.js","assets/CardList-BG11KU7H.js","assets/ItemList-D-9IbEjX.js","assets/Settings-CbgtqNcx.js","assets/settings-D-mx9vFJ.js","assets/Rules-DCuiI2LD.js","assets/Notifications-xaqRCmgU.js","assets/Chat-Di7cjGAh.js","assets/audio-codec-mjVr_wwJ.js"])))=>i.map(i=>d[i]);
+import{d as P,j as e,a as i,e as H}from"./react-vendor-Bf9ELcll.js";import{r as j,c as L,g as Z,a as Y}from"./chat-metadata-74T0w35U.js";import{L as D,a as A,b as K,U as W,A as J,c as Q,d as X,M as ee,e as te,B as se,f as ne,Z as ae,g as re,h as oe,G as I,i as ie,j as le,k as ce}from"./icons-vendor-BSp7PTZg.js";import"./vendor-BE-_xUHH.js";(function(){const s=document.createElement("link").relList;if(s&&s.supports&&s.supports("modulepreload"))return;for(const r of document.querySelectorAll('link[rel="modulepreload"]'))m(r);new MutationObserver(r=>{for(const l of r)if(l.type==="childList")for(const o of l.addedNodes)o.tagName==="LINK"&&o.rel==="modulepreload"&&m(o)}).observe(document,{childList:!0,subtree:!0});function n(r){const l={};return r.integrity&&(l.integrity=r.integrity),r.referrerPolicy&&(l.referrerPolicy=r.referrerPolicy),r.crossOrigin==="use-credentials"?l.credentials="include":r.crossOrigin==="anonymous"?l.credentials="omit":l.credentials="same-origin",l}function m(r){if(r.ep)return;r.ep=!0;const l=n(r);fetch(r.href,l)}})();class de extends P.Component{constructor(){super(...arguments),this.state={hasError:!1},this.handleReload=()=>{window.location.reload()}}static getDerivedStateFromError(){return{hasError:!0}}render(){const{hasError:s}=this.state,{children:n}=this.props;return s?e.jsx("main",{className:"flex min-h-screen items-center justify-center bg-canvas p-6 text-ink",children:e.jsxs("section",{className:"max-w-md text-center",children:[e.jsx("h1",{className:"text-xl font-semibold",children:"页面暂时无法显示"}),e.jsx("button",{className:"ios-btn-primary mt-6 px-4 py-2",type:"button",onClick:this.handleReload,children:"重新加载"})]})}):n}}const ue=async t=>{const s=await j(n=>L.POST("/api/v1/session/login",{body:t,signal:n}),{skipAuthLogout:!0});return{...s,token:s.token??void 0}},he=async t=>{const s=await j(n=>L.POST("/api/v1/session/initialize",{body:{password:t},signal:n}),{skipAuthLogout:!0});return{...s,token:s.token??void 0}},me=async t=>j(s=>L.GET("/api/v1/session",{signal:s}),t),fe=async()=>({success:!0,message:(await j(s=>L.POST("/api/v1/session/logout",{body:{},signal:s}))).message}),$=i.createContext(void 0),xe=({children:t})=>{const[s,n]=i.useState(!1),[m,r]=i.useState(!1),[l,o]=i.useState(!0),[a,d]=i.useState(!1),c=()=>{n(!1),r(!1)};i.useEffect(()=>{const p=new AbortController;let x=!0;me({signal:p.signal}).then(b=>{if(x){if((b==null?void 0:b.initialized)===!1){d(!0),c();return}d(!1),b!=null&&b.authenticated?(n(!0),r(b.is_admin===!0)):c()}}).catch(b=>{!x||b instanceof DOMException&&b.name==="AbortError"||(d(!1),c())}).finally(()=>{x&&o(!1)});const u=()=>c();return window.addEventListener("auth:logout",u),(()=>{x=!1,p.abort(),window.removeEventListener("auth:logout",u)})},[]);const h=async p=>{const x=await ue(p);return x.success&&(d(!1),n(!0),r(x.is_admin===!0)),x},g=async p=>{const x=await he(p);return x.success&&(d(!1),n(!0),r(x.is_admin===!0)),x},f=async()=>{try{await fe()}finally{c()}};return e.jsx($.Provider,{value:{checkingAuth:l,isLoggedIn:s,isAdmin:m,needsInit:a,signIn:h,initialize:g,signOut:f},children:t})},M=()=>{const t=i.useContext($);if(!t)throw new Error("useSession 必须在 SessionProvider 内使用");return t},z="ydisks.sidebar.v1",pe=()=>{try{return window.localStorage.getItem(z)==="collapsed"}catch{return!1}},ge=t=>{try{window.localStorage.setItem(z,t?"collapsed":"expanded")}catch{}},be=({className:t="w-full h-full text-white"})=>e.jsx("svg",{className:t,viewBox:"0 0 256 256","aria-hidden":"true",children:e.jsx("g",{stroke:"none",fill:"none",fillRule:"evenodd",children:e.jsx("path",{d:"M121.73,57.0003 C138.901,57.0003 154.897,65.5053 164.52,79.7493 C166.029,81.9823 165.441,85.0143 163.209,86.5233 C160.976,88.0323 157.944,87.4433 156.436,85.2113 C148.629,73.6553 135.655,66.7563 121.73,66.7563 C104.076789,66.7563 88.6201962,77.6596007 82.5581814,93.8184545 C83.9120803,93.7133631 85.2800264,93.66 86.6602,93.66 C111.537727,93.66 132.442611,110.997575 137.926382,134.224255 L162.5425,92.109 L162.615557,91.9879406 L162.756884,91.77367 L162.911804,91.5639102 L162.98657,91.4716759 C163.351845,91.025169 163.786059,90.6571744 164.264191,90.3738733 L164.550153,90.2178351 L164.729252,90.1315693 L164.989203,90.021723 L165.230265,89.9354653 L165.542393,89.8445739 C165.666021,89.8126706 165.790834,89.7857324 165.916565,89.7637431 L166.116517,89.7342499 C166.29505,89.7098142 166.475001,89.6957386 166.655596,89.6918808 L166.7539,89.693 C196.8909,89.693 221.4079,114.21 221.4079,144.347 C221.4079,174.482 196.8909,199 166.7539,199 L125.0869,199 C122.3919,199 120.2079,196.815 120.2079,194.122 C120.2079,191.429 122.3919,189.244 125.0869,189.244 L166.7539,189.244 C191.5109,189.244 211.6519,169.104 211.6519,144.347 C211.6519,120.513406 192.985349,100.957891 169.504046,99.5322548 L112.7745,196.584 C111.8995,198.08 110.2965,199 108.5635,199 L86.6602,199 C57.6182,199 33.9902,175.373 33.9902,146.33 C33.9902,122.475163 49.9316132,102.273061 71.7206298,95.817503 L71.9446339,94.967678 C78.0600632,72.5405324 98.33054,57.0003 121.73,57.0003 Z M86.6602,103.417 C62.9972,103.417 43.7462,122.667 43.7462,146.33 C43.7462,169.992 62.9972,189.244 86.6602,189.244 L105.7645,189.244 L129.869462,148.004782 C129.678439,147.482425 129.5742,146.918336 129.5742,146.33 C129.5742,122.667 110.3232,103.417 86.6602,103.417 Z",fill:"currentColor"})})}),U=({sizeClass:t="w-12 h-12",logoClassName:s="w-full h-full text-white"})=>e.jsxs("div",{className:"relative z-10 flex items-center justify-center",children:[e.jsx("div",{className:"absolute -inset-1 bg-blue-500/20 rounded-xl blur opacity-25"}),e.jsxs("div",{className:`relative ${t} rounded-xl flex items-center justify-center group-hover:scale-105 transition-transform duration-300`,children:[e.jsxs("svg",{className:"absolute inset-0 w-full h-full",viewBox:"0 0 120 120","aria-hidden":"true",children:[e.jsx("defs",{children:e.jsxs("linearGradient",{id:"login-squircle-gradient",x1:"0",y1:"0",x2:"1",y2:"1",children:[e.jsx("stop",{offset:"0%",stopColor:"rgb(var(--color-brand))"}),e.jsx("stop",{offset:"100%",stopColor:"rgb(var(--color-brand-light))"})]})}),e.jsx("path",{d:"M 114.00 60.00 L 113.83 71.48 L 113.31 78.16 L 112.44 83.67 L 111.22 88.46 L 109.66 92.71 L 107.75 96.49 L 105.48 99.87 L 102.86 102.86 L 99.87 105.48 L 96.49 107.75 L 92.71 109.66 L 88.46 111.22 L 83.67 112.44 L 78.16 113.31 L 71.48 113.83 L 60.00 114.00 L 48.52 113.83 L 41.84 113.31 L 36.33 112.44 L 31.54 111.22 L 27.29 109.66 L 23.51 107.75 L 20.13 105.48 L 17.14 102.86 L 14.52 99.87 L 12.25 96.49 L 10.34 92.71 L 8.78 88.46 L 7.56 83.67 L 6.69 78.16 L 6.17 71.48 L 6.00 60.00 L 6.17 48.52 L 6.69 41.84 L 7.56 36.33 L 8.78 31.54 L 10.34 27.29 L 12.25 23.51 L 14.52 20.13 L 17.14 17.14 L 20.13 14.52 L 23.51 12.25 L 27.29 10.34 L 31.54 8.78 L 36.33 7.56 L 41.84 6.69 L 48.52 6.17 L 60.00 6.00 L 71.48 6.17 L 78.16 6.69 L 83.67 7.56 L 88.46 8.78 L 92.71 10.34 L 96.49 12.25 L 99.87 14.52 L 102.86 17.14 L 105.48 20.13 L 107.75 23.51 L 109.66 27.29 L 111.22 31.54 L 112.44 36.33 L 113.31 41.84 L 113.83 48.52 Z",fill:"url(#login-squircle-gradient)"})]}),e.jsx("div",{className:"relative z-10 flex items-center justify-center w-full h-full",children:e.jsx(be,{className:s})})]})]}),ve=()=>{const{checkingAuth:t,needsInit:s,signIn:n,initialize:m}=M(),[r,l]=i.useState(""),[o,a]=i.useState(""),[d,c]=i.useState(""),[h,g]=i.useState(""),[f,p]=i.useState(!1),[x,u]=i.useState(""),b=async v=>{v.preventDefault(),p(!0),u("");try{const w=await n({username:r,password:o});w.success||u(w.message||"登录失败")}catch(w){u(w instanceof Error&&w.message||"登录失败")}finally{p(!1)}},C=async v=>{if(v.preventDefault(),u(""),d.length<8){u("密码至少需要 8 个字符");return}if(d!==h){u("两次输入的密码不一致");return}p(!0);try{const w=await m(d);if(!w.success){u(w.message||"初始化失败，请重试");return}c(""),g("")}catch(w){u(w instanceof Error&&w.message||"初始化失败，请重试")}finally{p(!1)}};if(t)return e.jsx("main",{className:"flex min-h-screen items-center justify-center bg-surface","aria-label":"正在校验会话",children:e.jsx(D,{className:"h-8 w-8 animate-spin text-brand"})});const q=s?"首次设置管理员密码":"欢迎回来",G=s?"设置完成后会自动进入系统，管理员账号为 admin。":"Ydisks闲鱼助手 · 自动发货与管家系统";return e.jsx("main",{className:"flex min-h-screen items-center justify-center bg-canvas p-4 font-sans",children:e.jsxs("section",{className:"w-full max-w-lg rounded-xl border border-white bg-white/80 p-8 shadow-panel backdrop-blur-3xl md:p-12",children:[e.jsxs("header",{className:"mb-8 text-center",children:[e.jsx("div",{className:"mx-auto mb-6 flex justify-center",children:e.jsx(U,{sizeClass:"h-24 w-24"})}),e.jsx("h1",{className:"text-3xl font-extrabold text-gray-900",children:q}),e.jsx("p",{className:"mt-2 font-medium text-gray-600",children:G})]}),s?e.jsxs("form",{className:"space-y-5",onSubmit:C,children:[e.jsxs("label",{className:"relative block",children:[e.jsx("span",{className:"sr-only",children:"设置管理员密码"}),e.jsx(A,{className:"absolute left-5 top-1/2 h-5 w-5 -translate-y-1/2 text-gray-400"}),e.jsx("input",{autoFocus:!0,className:"ios-input h-14 w-full rounded-xl py-4 pl-14 pr-6",onChange:(v=>c(v.target.value)),placeholder:"设置管理员密码（至少 8 个字符）",type:"password",value:d})]}),e.jsxs("label",{className:"relative block",children:[e.jsx("span",{className:"sr-only",children:"确认管理员密码"}),e.jsx(K,{className:"absolute left-5 top-1/2 h-5 w-5 -translate-y-1/2 text-gray-400"}),e.jsx("input",{className:"ios-input h-14 w-full rounded-xl py-4 pl-14 pr-6",onChange:(v=>g(v.target.value)),placeholder:"再次输入密码",type:"password",value:h})]}),e.jsx(R,{submitting:f,text:"设置密码并进入系统"})]}):e.jsxs("form",{className:"space-y-5",onSubmit:b,children:[e.jsxs("label",{className:"relative block",children:[e.jsx("span",{className:"sr-only",children:"管理员账号"}),e.jsx(W,{className:"absolute left-5 top-1/2 h-5 w-5 -translate-y-1/2 text-gray-400"}),e.jsx("input",{autoFocus:!0,className:"ios-input h-14 w-full rounded-xl py-4 pl-14 pr-6",onChange:(v=>l(v.target.value)),placeholder:"管理员账号",type:"text",value:r})]}),e.jsxs("label",{className:"relative block",children:[e.jsx("span",{className:"sr-only",children:"管理员密码"}),e.jsx(A,{className:"absolute left-5 top-1/2 h-5 w-5 -translate-y-1/2 text-gray-400"}),e.jsx("input",{className:"ios-input h-14 w-full rounded-xl py-4 pl-14 pr-6",onChange:(v=>a(v.target.value)),placeholder:"密码",type:"password",value:o})]}),e.jsx(R,{submitting:f,text:"立即登录"})]}),x?e.jsx("p",{className:"mt-5 rounded-xl bg-red-50 p-3 text-center text-sm font-medium text-red-600",children:x}):null]})})},R=({submitting:t,text:s})=>e.jsx("button",{className:"ios-btn-primary flex h-14 w-full items-center justify-center gap-2 rounded-xl text-lg disabled:opacity-70",disabled:t,type:"submit",children:t?e.jsx(D,{className:"h-5 w-5 animate-spin"}):e.jsxs(e.Fragment,{children:[e.jsx("span",{children:s}),e.jsx(J,{className:"h-5 w-5"})]})}),we="modulepreload",ye=function(t){return"/static/"+t},T={},y=function(s,n,m){let r=Promise.resolve();if(n&&n.length>0){let o=function(c){return Promise.all(c.map(h=>Promise.resolve(h).then(g=>({status:"fulfilled",value:g}),g=>({status:"rejected",reason:g}))))};document.getElementsByTagName("link");const a=document.querySelector("meta[property=csp-nonce]"),d=(a==null?void 0:a.nonce)||(a==null?void 0:a.getAttribute("nonce"));r=o(n.map(c=>{if(c=ye(c),c in T)return;T[c]=!0;const h=c.endsWith(".css"),g=h?'[rel="stylesheet"]':"";if(document.querySelector(`link[href="${c}"]${g}`))return;const f=document.createElement("link");if(f.rel=h?"stylesheet":we,h||(f.as="script"),f.crossOrigin="",f.href=c,d&&f.setAttribute("nonce",d),document.head.appendChild(f),h)return new Promise((p,x)=>{f.addEventListener("load",p),f.addEventListener("error",()=>x(new Error(`Unable to preload CSS for ${c}`)))})}))}function l(o){const a=new Event("vite:preloadError",{cancelable:!0});if(a.payload=o,window.dispatchEvent(a),!a.defaultPrevented)throw o}return r.then(o=>{for(const a of o||[])a.status==="rejected"&&l(a.reason);return s().catch(l)})},je=({activeTab:t,isAdmin:s=!1,collapsed:n,onToggleCollapsed:m,onNavigate:r,onLogout:l,buildInfo:o,hasUnreadChatMessage:a=!1})=>{const d=[{id:"dashboard",icon:Q,label:"仪表盘"},{id:"accounts",icon:X,label:"账号管理"},{id:"chat",icon:ee,label:"在线聊天"},{id:"cards",icon:te,label:"卡密库存"},{id:"items",icon:se,label:"商品列表"},{id:"orders",icon:ne,label:"订单管理"},{id:"rules",icon:ae,label:"自动化规则"},{id:"notifications",icon:re,label:"通知设置"},...s?[{id:"settings",icon:oe,label:"系统与AI"}]:[]],c=/^\d+\.\d+\.\d+$/.test(o.version)?`v${o.version}`:o.version;return e.jsxs("aside",{className:`fixed inset-y-0 left-0 z-20 flex flex-col border-r border-slate-200/80 bg-white/95 shadow-sidebar backdrop-blur-xl transition-[width] duration-300 ${n?"w-16":"w-64"}`,children:[e.jsxs("div",{className:`flex h-20 items-center border-b border-slate-100 ${n?"justify-center px-2":"gap-3 px-5"}`,children:[e.jsx(U,{sizeClass:"h-10 w-10"}),!n&&e.jsxs("div",{className:"min-w-0 leading-tight",children:[e.jsx("div",{className:"truncate text-base font-black tracking-tight text-slate-950",children:"Ydisks 闲鱼助手"}),e.jsx("div",{className:"mt-1 text-[10px] font-extrabold uppercase tracking-[0.22em] text-sky-600",children:"Operations"})]})]}),e.jsx("nav",{className:`flex-1 space-y-1.5 overflow-y-auto pt-5 ${n?"px-2":"px-3"}`,"aria-label":"主导航",children:d.map(h=>{const g=h.icon,f=t===h.id;return e.jsx(P.Fragment,{children:e.jsxs("button",{type:"button",title:n?h.label:void 0,"aria-label":h.label,"aria-current":f?"page":void 0,onClick:(()=>r(h.id)),className:`group relative flex h-11 w-full items-center rounded-xl transition-colors ${n?"justify-center":"gap-3 px-3.5"} ${f?"bg-brand text-white shadow-brand-active":"text-slate-500 hover:bg-slate-100 hover:text-slate-900"}`,children:[e.jsxs("span",{className:"relative flex h-[19px] w-[19px] shrink-0",children:[e.jsx(g,{className:`h-[19px] w-[19px] ${f?"text-white":"text-slate-400 group-hover:text-slate-700"}`}),h.id==="chat"&&a&&e.jsx("span",{className:"absolute -right-1 -top-1 h-3 w-3 rounded-full border-2 border-white bg-red-500",role:"status","aria-label":"在线聊天有未读消息"})]}),!n&&e.jsx("span",{className:"truncate text-sm font-bold",children:h.label}),f&&!n&&e.jsx("span",{className:"ml-auto h-1.5 w-1.5 rounded-full bg-white/90"})]})},h.id)})}),e.jsxs("div",{children:[e.jsx("div",{title:n?`版本 ${o.version} · ${o.commit}`:void 0,className:`border-y border-slate-100 bg-slate-50/70 py-2.5 ${n?"flex justify-center px-1":"px-6"}`,children:n?e.jsx(I,{className:"h-[18px] w-[18px] text-slate-400","aria-label":`版本 ${o.version}`}):e.jsxs("div",{className:"min-w-0",children:[e.jsxs("div",{className:"flex items-center gap-2 text-[10px] font-extrabold uppercase tracking-[0.18em] text-slate-400",children:[e.jsx(I,{className:"h-3.5 w-3.5"}),"运行版本"]}),e.jsxs("div",{className:"mt-1 flex items-baseline gap-2",children:[e.jsx("span",{className:"truncate font-mono text-xs font-bold text-slate-700",children:c}),e.jsx("span",{className:"truncate font-mono text-[10px] text-slate-400",children:o.commit})]})]})}),e.jsxs("div",{className:`space-y-2 p-2 ${n?"":"p-3"}`,children:[e.jsxs("button",{type:"button",onClick:m,title:n?"展开侧边栏":"收起侧边栏","aria-label":n?"展开侧边栏":"收起侧边栏",className:`flex h-10 w-full items-center rounded-xl text-slate-500 transition-colors hover:bg-slate-100 hover:text-slate-900 ${n?"justify-center":"gap-3 px-3"}`,children:[n?e.jsx(ie,{className:"h-5 w-5"}):e.jsx(le,{className:"h-5 w-5"}),!n&&e.jsx("span",{className:"text-sm font-bold",children:"收起侧边栏"})]}),e.jsxs("button",{type:"button",onClick:l,title:n?"退出登录":void 0,"aria-label":"退出登录",className:`flex h-10 w-full items-center rounded-xl text-slate-500 transition-colors hover:bg-red-50 hover:text-red-600 ${n?"justify-center":"gap-3 px-3"}`,children:[e.jsx(ce,{className:"h-5 w-5"}),!n&&e.jsx("span",{className:"text-sm font-bold",children:"退出登录"})]})]})]})]})},S=new Set;let B="connecting";const k=new Set,Ye=t=>(t({type:"connection",state:B}),S.add(t),(()=>S.delete(t))),N=t=>{B=t;for(const s of S)s({type:"connection",state:t})},Le=t=>{for(const s of S)s({type:"message",message:t})},Ce=t=>(k.add(t),(()=>k.delete(t))),Ke=t=>{for(const s of k)s(t)},Ne=1e3,Se=t=>`【新消息】${t}`,Ee=()=>{const[t,s]=i.useState(0),[n,m]=i.useState(!1),r=i.useRef(document.title),l=i.useRef(0);return i.useEffect(()=>{const a=()=>s(0),d=()=>{document.visibilityState==="visible"&&a()};return window.addEventListener("focus",a),document.addEventListener("visibilitychange",d),(()=>{window.removeEventListener("focus",a),document.removeEventListener("visibilitychange",d)})},[]),i.useEffect(()=>{if(t<=0){document.title=r.current;return}let a=!0;const d=()=>{document.title=a?Se(r.current):r.current};d();const c=window.setInterval(()=>{a=!a,d()},Ne);return(()=>{window.clearInterval(c),document.title=r.current})},[t]),i.useEffect(()=>{const a=new AbortController,d=l.current;return(async()=>{try{const g=(await Z({signal:a.signal})).filter(x=>x.enabled),f=await Promise.all(g.map(x=>Y(x.id,void 0,{signal:a.signal})));if(a.signal.aborted||l.current!==d)return;const p=f.some(x=>x.sessions.some(u=>u.unread_count>0));m(p)}catch{}})(),(()=>a.abort())},[]),i.useEffect(()=>Ce(m),[]),{notifyIncomingMessage:i.useCallback(a=>{a.direction!=="incoming"||a.message_type==="system"||(l.current+=1,s(d=>d+1),m(!0))},[]),hasUnreadChatMessage:n}},_e=()=>{const t=Ee(),{notifyIncomingMessage:s,hasUnreadChatMessage:n}=t;return i.useEffect(()=>{let m=!1,r=0,l=0,o=null;const a=()=>{if(m)return;N("connecting");const d=window.location.protocol==="https:"?"wss:":"ws:";o=new WebSocket(`${d}//${window.location.host}/api/v1/chat/ws`),o.onopen=()=>{l=0,N("online")},o.onmessage=c=>{try{const g=JSON.parse(c.data).message;g&&(s(g),Le(g))}catch{}},o.onclose=()=>{if(m)return;N("offline");const c=Math.min(15e3,1e3*2**Math.min(l++,4));r=window.setTimeout(a,c)},o.onerror=()=>o==null?void 0:o.close()};return a(),(()=>{m=!0,window.clearTimeout(r),o==null||o.close(),N("offline")})},[s]),{hasUnreadChatMessage:n}},ke=async t=>j(s=>L.GET("/health",{signal:s}),t),E=i.lazy(()=>y(()=>import("./Dashboard-DBUkwLYQ.js"),__vite__mapDeps([0,1,2,3,4,5,6,7,8]))),Pe=i.lazy(()=>y(()=>import("./AccountList-D1PDvKjr.js"),__vite__mapDeps([9,1,7,5,6,3,8]))),Ae=i.lazy(()=>y(()=>import("./OrderList-CKY0um79.js"),__vite__mapDeps([10,1,2,3,7,5,6,8]))),Ie=i.lazy(()=>y(()=>import("./CardList-BG11KU7H.js"),__vite__mapDeps([11,1,7,5,6,3,8]))),Re=i.lazy(()=>y(()=>import("./ItemList-D-9IbEjX.js"),__vite__mapDeps([12,1,7,5,6,3,8]))),Te=i.lazy(()=>y(()=>import("./Settings-CbgtqNcx.js"),__vite__mapDeps([13,1,7,5,6,3,14]))),Oe=i.lazy(()=>y(()=>import("./Rules-DCuiI2LD.js"),__vite__mapDeps([15,1,3,7,5,6,8]))),De=i.lazy(()=>y(()=>import("./Notifications-xaqRCmgU.js"),__vite__mapDeps([16,1,3,14,7,5,6]))),$e=i.lazy(()=>y(()=>import("./Chat-Di7cjGAh.js"),__vite__mapDeps([17,1,5,6,18,3,7]))),Me=()=>e.jsx("div",{className:"flex min-h-[24rem] items-center justify-center",role:"status","aria-label":"正在加载页面",children:e.jsx("div",{className:"h-8 w-8 animate-spin rounded-full border-4 border-brand/20 border-t-brand"})}),ze=({activeTab:t,isAdmin:s,deliveryRuleTarget:n,onConfigureDelivery:m,onDeliveryTargetHandled:r})=>{const l=a=>{m({cookieId:a.cookie_id,itemId:a.item_id,requestId:Date.now()})},o=()=>{switch(t){case"dashboard":return e.jsx(E,{});case"accounts":return e.jsx(Pe,{});case"chat":return e.jsx($e,{});case"orders":return e.jsx(Ae,{});case"cards":return e.jsx(Ie,{});case"items":return e.jsx(Re,{onConfigureDelivery:l});case"rules":return e.jsx(Oe,{initialDeliveryTarget:n,onDeliveryTargetHandled:r});case"notifications":return e.jsx(De,{isAdmin:s});case"settings":return s?e.jsx(Te,{}):e.jsx(E,{});default:return e.jsx(E,{})}};return e.jsx(i.Suspense,{fallback:e.jsx(Me,{}),children:o()})},Ue=({activeTab:t,isAdmin:s,collapsed:n,deliveryRuleTarget:m,onToggleCollapsed:r,onNavigate:l,onLogout:o,onConfigureDelivery:a,onDeliveryTargetHandled:d})=>{const[c,h]=i.useState({version:"dev",commit:"unknown"}),{hasUnreadChatMessage:g}=_e();return i.useEffect(()=>{const f=new AbortController;return ke({signal:f.signal}).then(p=>h({version:String(p.version||"dev"),commit:String(p.commit||"unknown")})).catch(p=>{p instanceof DOMException&&p.name}),(()=>f.abort())},[]),e.jsxs("div",{className:"flex min-h-screen bg-canvas text-ink",children:[e.jsx(je,{activeTab:t,isAdmin:s,collapsed:n,onToggleCollapsed:r,onNavigate:l,onLogout:o,buildInfo:c,hasUnreadChatMessage:g}),e.jsxs("main",{className:`h-screen min-w-0 flex-1 overflow-x-hidden overflow-y-auto scroll-smooth transition-[margin] duration-300 ${n?"ml-16":"ml-64"} ${t==="chat"?"p-4 md:p-6":"p-8 md:p-12"}`,children:[e.jsx("div",{className:"fixed top-0 right-0 -z-10 h-[800px] w-[800px] pointer-events-none rounded-full bg-gradient-to-bl from-blue-50 to-transparent opacity-60 blur-[120px]"}),e.jsx("div",{className:`${t==="chat"?"mx-auto max-w-[1680px]":"mx-auto max-w-[1400px] pb-10"}`,children:e.jsx(ze,{activeTab:t,isAdmin:s,deliveryRuleTarget:m,onConfigureDelivery:a,onDeliveryTargetHandled:d})})]})]})},V={"/app/dashboard":"dashboard","/app/accounts":"accounts","/app/chat":"chat","/app/orders":"orders","/app/cards":"cards","/app/items":"items","/app/rules":"rules","/app/notifications":"notifications","/app/settings":"settings"},_=Object.fromEntries(Object.entries(V).map(([t,s])=>[s,t])),O=()=>V[window.location.pathname]??"dashboard",Be=()=>{const{isLoggedIn:t,isAdmin:s,signOut:n}=M(),[m,r]=i.useState(O),[l,o]=i.useState(),[a,d]=i.useState(pe);i.useEffect(()=>{const u=()=>r(O());return window.addEventListener("popstate",u),(()=>window.removeEventListener("popstate",u))},[]),i.useEffect(()=>{t&&!s&&m==="settings"&&(window.history.replaceState({},"",_.dashboard),r("dashboard"))},[m,s,t]);const c=u=>{const b=u==="settings"&&!s?"dashboard":u,C=_[b];C!==window.location.pathname&&window.history.pushState({},"",C),r(b)},h=u=>{const b=Object.prototype.hasOwnProperty.call(_,u)?u:"dashboard";c(b)},g=async()=>{try{await n()}catch(u){console.error("退出登录失败",u)}},f=()=>{d(u=>{const b=!u;return ge(b),b})},p=u=>{o(u),c("rules")},x=()=>o(void 0);return t?e.jsx(Ue,{activeTab:m,isAdmin:s,collapsed:a,deliveryRuleTarget:l,onToggleCollapsed:f,onNavigate:h,onLogout:g,onConfigureDelivery:p,onDeliveryTargetHandled:x}):e.jsx(ve,{})},Ve=()=>e.jsx(de,{children:e.jsx(xe,{children:e.jsx(Be,{})})}),F=document.getElementById("root");if(!F)throw new Error("Could not find root element to mount to");const Fe=H.createRoot(F);Fe.render(e.jsx(P.StrictMode,{children:e.jsx(Ve,{})}));export{Ke as p,Ye as s};
