+++ internal/webui/static/assets/settings-D-mx9vFJ.js
@@ -1,0 +1,1 @@
+const a=new Set(["ai_api_key","smtp_password","qq_reply_secret_key","captcha.remote_secret_key"]),c=t=>{if("values"in t||"secrets"in t)return t;const n={},o={};for(const[r,e]of Object.entries(t))if(e!=null){if(a.has(r)){o[r]=e===""?{action:"clear"}:{action:"replace",value:String(e)};continue}(typeof e=="string"||typeof e=="number"||typeof e=="boolean")&&(n[r]=e)}return Object.keys(o).length>0?{values:n,secrets:o}:n};export{a as S,c as n};
