+++ internal/webui/static/index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Ydisks闲鱼助手 - 自动发货与管家系统</title>
     <meta name="description" content="Ydisks闲鱼助手 - 自动回复、自动发货、订单管理、AI智能议价一体化解决方案" />
-    <script type="module" crossorigin src="/static/assets/index-dO7JyNQY.js"></script>
+    <script type="module" crossorigin src="/static/assets/index-Q3nlx6gR.js"></script>
     <link rel="modulepreload" crossorigin href="/static/assets/react-vendor-Bf9ELcll.js">
     <link rel="modulepreload" crossorigin href="/static/assets/vendor-BE-_xUHH.js">
     <link rel="modulepreload" crossorigin href="/static/assets/icons-vendor-BSp7PTZg.js">
